+++ frontend/src/components/Card/Card.js
@@ -1,68 +1,20 @@
-import Component from '../component';
+import CustomComponent from '../CustomComponent';
 import viewImage from '../../assets/icons/view.svg';
 import bookmarkImage from '../../assets/icons/bookmark.svg';
-import bookmarkFilledImage from '../../assets/icons/bookmark_filled.svg';
 import PostBanner from '../PostBanner/PostBanner';
-import './card.scss';
-import axiosInstance from '../../utils/api';
+import './Card.scss';
 import RouterContext from '../../router/RouterContext';
-import Toast from '../Toast/Toast';
 import { shortSido } from '../../library/MainPage/index';
-import CustomComponent from '../../components/CustomComponent';
-import { createDom } from '../../utils/dom';
-
-export default class Card extends CustomComponent {
+import { createDom, replaceElement } from '../../utils/dom';
+export default class MainCard extends CustomComponent {
   init() {
-    // super(props);
-
     this.state = {
       type: this.props.type,
-      post: props.post,
-      postList: props.postList,
-      updateCards: props.updateCards,
+      post: this.props.post,
     };
-
-    this.$dom = this.createDom('div', {
-      className: 'card-wrapper',
-    });
   }
 
-  renderCallback() {
-    return `
-    <div class="card-wrap">
-        <div class="image"></div>
-        <div class="card-body">
-            <div class="card-title">${this.state.post.title}</div>
-            <div class="card-info-wrapper">
-                <div class="card-info">
-                    <div class="card-info-detail">
-                        <div class="text">${this.makeSido(
-                          this.state.post.address,
-                        )}</div>
-                    </div>
-                    <div class="card-info-detail">
-                        <div class="text">${this.state.post.capacity}명</div>
-                    </div>
-                </div>
-                <div class="card-info-number">
-                    <div class="card-info-number-detail">
-                        <img src="${viewImage}" aria-label="조회수"/>
-                        <div>${this.state.post.views}</div>
-                    </div>
-                    <div class="card-info-number-detail">
-                        <button type="button" class="bookmark-btn" aria-label="북마크">
-                          <img src="${bookmarkImage}" aria-label="북마크"/>
-                        </button>
-                        <div>${this.state.post.marks}</div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-    `;
-  }
-
-  makeSido = address => {
+  makeSido(address) {
     if (address.split(' ')[0] === '온라인') {
       return '온라인';
     } else if (address.split(' ')[0] === '세종특별자치시') {
@@ -73,53 +25,60 @@ export default class Card extends CustomComponent {
       }`;
     }
     return `${address.split(' ')[0]} ${this.state.post.address.split(' ')[1]}`;
-  };
+  }
+
+  markup() {
+    return (
+      <div class="card">
+        <div class="card__image"></div>
+        <div class="card__body">
+          <div class="card__body__title">{this.state.post.title}</div>
+          <div class="card__body__info">
+            <div class="card__body__info__content">
+              <div class="card__body__info__content__detail">
+                <div class="card__body__info__content__detail__text">
+                  {this.makeSido(this.state.post.address)}
+                </div>
+              </div>
+              <div class="card__body__info__content__detail">
+                <div class="card__body__info__content__detail__text">
+                  {this.state.post.capacity}명
+                </div>
+              </div>
+            </div>
+            <div class="card__body__info__number">
+              <div class="card__body__info__number__detail">
+                <img src={viewImage} aria-label="조회수" />
+                <div>{this.state.post.views}</div>
+              </div>
+              <div class="card__body__info__number__detail">
+                <img src={bookmarkImage} aria-label="북마크" />
+                <div>{this.state.post.marks}</div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    );
+  }
 
   setEvent() {
-    const images = this.$dom.querySelector('.image');
-    const cardWrap = this.$dom.querySelector('.card-wrap');
-    const bookmarkBtn = this.$dom.querySelector('.bookmark-btn');
-    const bookmark = this.$dom.querySelector('.bookmark-btn').childNodes[1];
+    const images = this.container.querySelector('.card__image');
+    const card = this.container.querySelector('.card');
 
-    this.replaceElement(
-      images,
-      new PostBanner({ stackList: this.state.post.stacks }).$dom,
-    );
-
-    cardWrap.addEventListener('click', event => {
-      if (event.target !== bookmark) {
-        RouterContext.state.push(`/detail/${this.state.post._id}`);
-      }
-      if (this.state.type === 'bookmark') {
-        this.state.updateCards();
-      }
+    const div = createDom('div', {
+      className: `banner banner--${this.state.post.stacks[0]}`,
     });
 
-    if (this.state.type === 'bookmark') {
-      bookmark.src = bookmarkFilledImage;
-      bookmarkBtn.addEventListener('click', () => {
-        bookmarkPost();
-      });
-    } else {
-      bookmarkBtn.className = 'bookmark-btn-main';
-    }
+    new PostBanner({
+      container: div,
+      props: { stacks: this.state.post.stacks },
+    });
 
-    const bookmarkPost = () => {
-      for (let i = 0; i < this.state.postList.length; i++) {
-        // 이미 북마크한 포스트의 북마크 버튼을 누른 경우 북마크를 해제한다.
-        if (this.state.post._id === this.state.postList[i]._id) {
-          axiosInstance.delete(`/users/mark/${this.state.post._id}`, {
-            withCredentials: true,
-          });
-          break;
-        }
-      }
-      this.state.updateCards();
-      new Toast({
-        timeout: 2100,
-        content: '북마크가 해제되었습니다',
-        type: 'success',
-      });
-    };
+    replaceElement(images, div);
+
+    card.addEventListener('click', () => {
+      RouterContext.state.push(`/detail/${this.state.post._id}`);
+    });
   }
 }
