--- frontend/src/components/MainCard/MainCard.scss
@@ -1,117 +1,0 @@
-.card-wrapper {
-  display: flex;
-  flex-direction: column;
-  background-color: var(--white);
-  width: 280px;
-  height: 320px;
-  margin: 60px;
-  border-radius: 12px;
-  filter: drop-shadow(4px 4px 12px rgba(0, 0, 0, 0.25));
-}
-.card-wrap {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
-
-  .banner {
-    flex-basis: 50%;
-    display: flex;
-    align-items: center;
-    flex-direction: row;
-    flex-wrap: wrap;
-    overflow: hidden;
-    border-top-left-radius: 12px;
-    border-top-right-radius: 12px;
-    .bannerLogo {
-      width: 42px;
-      height: 42px;
-      border-radius: 4px;
-      filter: drop-shadow(4px 4px 4px rgba(0, 0, 0, 0.25));
-    }
-  }
-}
-
-.card-body {
-  height: 50%;
-}
-
-.card-title {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  font-size: 16px;
-  font-weight: bold;
-  text-align: center;
-  line-height: 48px;
-  padding-left: 30px;
-  padding-right: 30px;
-  display: -webkit-box;
-  -webkit-line-clamp: 1;
-  -webkit-box-orient: vertical;
-  word-wrap: break-word;
-}
-
-.card-info-wrapper {
-  display: grid;
-  grid-template-columns: 52% 48%;
-  align-items: center;
-  // padding-bottom: 10px;
-}
-
-.card-info-detail {
-  width: 100px;
-  height: 34px;
-  border-radius: 30px;
-  background-color: #f0f0f0;
-  margin-top: 15px;
-  margin-left: 15px;
-}
-
-.text {
-  text-align: center;
-  transform: translateY(70%);
-  font-size: 14px;
-}
-
-.card-info-number {
-  display: grid;
-  grid-template-columns: 50% 50%;
-  margin-top: 63px;
-  margin-right: 20px;
-}
-
-.card-info-number-detail {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 8px;
-}
-
-.card-info-number-detail img {
-  padding-right: 8px;
-  width: 30px;
-  height: 30px;
-}
-
-.bookmark-btn,
-.bookmark-btn-main {
-  width: 30px;
-  height: 30px;
-  background-color: transparent;
-  border: 0;
-}
-
-.card-wrapper:hover {
-  transform: scale(1.05);
-  transition: transform 0.2s, opacity 0.2s, background-color 0.2s;
-  cursor: pointer;
-}
-
-.bookmark-btn:hover {
-  cursor: pointer;
-}
-
-.bookmark-btn-main:hover {
-  transform: scale(1);
-  cursor: pointer;
-}
